Selects slices for manual labeling - Issue #1
Adds to the list of slice selections in segmentation_slices.csv. This
file contains a list of slice numbers for each brain noting the most
anterior coronal slice of the putamen, the most posterior coronal slice
of the thalamus and a selection slice. This selection slice is chosen
to have interesting features, usaully involving the visibility of at
least the thalamus, palidum, putamen and hippocampus.

Slice numbers are given in relation to the order of slices in the hard
reconstructed volume. In this ordering the most anterior slice of the
brain is slice number 3. This is due to the two padding slices, used to
enforce out of plane affine adjustment, as well as conversion to the 1
ordinal counting system of matlab (as opposed the the zero ordinal
system in freeview).

As this system refers to the slice order in the volume adjustments have
to be made during look up, to account for the reordering of slices in
some cases.