Adds script to apply MRI seg to photos - Issue #1
Adds the script UWphoto_script_mriSegPhotoOverlay.m to transfer the
Samseg segmentations of the MRI images on to the hard registered photo
volumes. This is initially used to pick the slices for manual
segmentation but may later be used for comparison of the segmentations.

The script works by taking the header information from the warped
reference, generated during the hard photo reconstruction, and applying
them to the pixel data of the MRI segmentations.

Currently the script is only set up for the file structure on the linux
machine for H Tregidgo. This will require photo and segmentation
directories to be changed in lines 7 & 8 for others to use.