Fixes photo flipping in soft reconstruction
Fixes the problem of a left right flip occuring during the volume
reconstruction process. We found that features in some asymmetric
brains were flipped in the MRI's vs the photo volumes. Looking further
the photo volumes appeared flipped vs their source images.

The fix was to modify the initial vox2ras matrix with a negative value
in the diagonal element corresponding to that voxel dimension. The fix
was initially implemented for the hard reconstruction and has been
added here to the soft reconstruction code.