Adds volume code for soft manual labels - Issue #4
Adds a cost function, volume recon function and label processing script
to place the manual labels in the geometry of the soft reconstructed
photo volumes.

The cost function is a modified version of the cost function used in
the reconstruction of the volumes as this function is capable of taking
saved optimisation parameters and building a resampled volume. The only
modification is a switch to using nearest neighbour interpolation due
to the discrete nature of the labels.

The reconstruction function is a modified version of the soft
reconstruction function. It is modified to identify the correct slice
that has been labeled (from the string format of the file name), insert
the labels into the first colour channel of the photo slice and run the
cost function with saved parameters. This function also requires
nearest neighbour resampling as well as care to match the array types
(png data comes in as uint8 and can cause overflow with some labels,
conversion to double requires rounding when outputing).

Note a change in the filename format for the manual labels will result
in this function breaking.