Adds matlab & bash settup code - Issue #2
Adds the script UWphoto_startup.m to help with adding the correct paths
and environment variables to allow other functions to work. In
particular this script adds the photo processing code base, checks that
the freesurfer matlab and binary paths are added to the bash path and
initialises the FREESURFER_HOME and FS_PYTHON_DEV environment variables
for use with python code. Additionally this code sets up a
PHOTO_RECON_HOME environment variable. This makes note of the top level
directory so that relative paths in the filestructure can be coded.

This commit also adds UWphoto_function_checkBashEnvironment.m which
sets up the FREESURFER_HOME and FS_PYTHON_DEV environment variables for
use with python code modified from Samseg. This has been set up so that
the compiled python packages and interpreter from a stable or dev build
of freesurfer can be used to call code from a dev directory. This
allows code in a git repo to be called without having to be compiled or
sorting out a virtual environment. (this is a little bit hackey and may
be modified later).

At the moment these two files are set up with some directories hard
coded to work on one specific machine. This is marginally better than
before as, once all the code is modified to use the PHOTO_RECON_HOME
environment variable, paths should only need to be modified in two
places rather than in every file. However, as modifications to these
files may not play nicely with git, eventually we may want to remove
them from the repository and instead have an "Initial setup" script/
function to auto generate them on any additional machines.