Fixes label import masking - Issue #4
Modifies UWphoto_function_segFromManualLabel.m to prevent labels from
being incorrectly masked out. This was due to an issue where the photo
mask, used for aligning centres of gravity (COG) in the registration
process, was applied to the photo data before the optimisation process.

While for most photos a imfill('holes') step prevented the ventricals
from being masked out, some had deformationa so severe that the
ventricals were connected to the background. This meant that labels for
ventricals and other places where the structure was connected to the
background were being masked out. The solution was to prevent the
masking step in the re-import code.

A flag has also been added to the re-import script to manual force re-
generation of already complete label volumes when necessary.