Adds correlation calculation and plots - Issue #8
Adds the script UWpoto_script_correlationGraphs.m to plot scatter
graphs of Photo label volumes vs MRI label volumes for each of the
cases. Pearson's correlation coeficients (r) and p values are also
calculated for each of the plots and stored in the structure
CorrelationStruct.

CorrelationStruct holds the raw volumes for each volume and
segmentation type as well as the correlation statistics and  results of
a robustfit operation to generate a trend line. This robustfit is
performed with the function present in Matlab's statistics and machine
learning toolbox and should account for outliers.

Outstanding issues from this script are some extreme values which may
be biasing without reaching the threshold to be labeled as an outlier
by robustfit. Additionally the code to perform GLMs on the data in
combination with the case information needs to be checked and uploaded.