Adds handling for excluded cases - Issue #9
Modifies the scripts UWphoto_script_extractStats.m and
UWphoto_script_correlationGraphs.m to handle excluded cases.

The modified stat extraction code takes a list of case IDs to be
removed from the analysis and asigns a removal level flag, with notes
on the case, to the output volume structure segVolumeInfo. These
removal flags can then be used to prioritise which cases to be removed
before generating the volume scatter plots and calculating
correlations.

Level 1 relates to the cases with missing slices or including large
hemmorhage regions. Level 2 relates to those with extreme problems with
the lateral ventricles, as seen on MRI. Level 3 relates to cases with
poor MRI segmentations or other deformations.

A possible level 4 may be added to hold cases which have no issues with
reconstruction or segmentation but which have extreme values that may
bias the correlation statistics.