Fixes movie generation for paper figure - Issue #12
Modifies the script makeMovieFramesHardMask.m to account for the
modifications made to the hard reconstruction algorithm. These include
the addition of frames either side of the volume, enforcing stretch in
the anterior-posterior direction, and the re-ordering of slices where
appropriate.

Volumes generated by this script have been used in making figures for
the paper.