Adds handling for single hemispheres - Issue #14
Modifies the script UWphoto_script_SoftReconstructions.m to remove hard-
coded paths to data and results directories. Also changes the atlas for
reconstruction to the left hemisphere atlas. This script is now set up
for one specific run and still needs modification to allow general use.

The function ReconPhotoVolume_joint_multires.m has also been modified.
In particular the code now initialises the translations for each slice
to match the centre of gravity of corresponding slices on the roughly
aligned reference image.

Additionally a bug has been fixed that was causing rotational and
affine transformations to be applied using the wrong coordinates for
the centre of gravity.