Adds selection for hemisphere and view - Issue #14
Modifies the script UWphoto_script_SoftReconstructions.m to allow both
the hemisphere and viewpoint of the photos to be selected by the user.
The script assumes all cases within the top level source directory have
the same viewpoint and hemisphere. The user is presented with question
boxes asking to select left, right or both hemispheres and then asked
if the slices are anterior or posterior views.

Additionally a bug is fixed in the input parsing for the APswitch
parameter for function ReconPhotoVolume_joint_multires.m.