Add previously untracked SAMSEG files - Issue #6
This commit adds all the untracked SAMSEG scripts and functions to git.
Some of these were used in generating the MICCAI paper but were not
added to git at the time due to time constraints and also as we hadn't
decided whether to keep them in the same repo as the reconstruction code
or add them to the repo being used for the photo version of SAMSEG.
These files really need to be in source control now so I am adding them
here.

Initially these files are being added to the branch
	seg-dev-at-submission
to indicate some of these files are required to run the experiments for
the MICCAI paper.

Another branch
	seg-dev-current
will be added to track the ongoing segmentation code development and
allow code from branches created after MICCAI submission to be merged in.